+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## CHANGE LOG
 
+### v5.0.0
+
+2013-06-11 issue [#62](https://github.com/qiniu/api/pull/62)
+
+- 遵循 [sdkspec v1.0.2](https://github.com/qiniu/sdkspec/tree/v1.0.2)
+  - rs.GetPolicy 删除 Scope，也就是不再支持批量下载的授权。
+  - rs.New, PutPolicy.Token, GetPolicy.MakeRequest 增加 mac *digest.Mac 参数。
+- 初步整理了 sdk 使用文档。
+
+
 ### v0.9.1
 
 2013-05-28 issue [#56](https://github.com/qiniu/api/pull/56)